refactor: Refatora o controller para incluir o ResponseEntity nas mihas rotas

+++ CadastroDeNinjas/src/main/java/dev/java10x/CadastroDeNinjas/Ninjas/NinjaController.java
@@ -1,7 +1,7 @@
 package dev.java10x.CadastroDeNinjas.Ninjas;
-
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
 
 @RestController
@@ -20,28 +20,54 @@ public class NinjaController {
     }
 
     @PostMapping("/criar")
-    public NinjaDTO criarNinja(@RequestBody NinjaDTO ninja) {
-        return ninjaService.criarNinja(ninja);
+    public ResponseEntity<String> criarNinja(@RequestBody NinjaDTO ninja) {
+
+        NinjaDTO novoNinja = ninjaService.criarNinja(ninja);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Ninja criado com sucesso: " + novoNinja.getNome() + "(ID): " + novoNinja.getId());
+
     }
 
     @GetMapping("/listar")
-    public List<NinjaDTO> listarNinjas() {
-        return ninjaService.listarNinjas();
+    public ResponseEntity<List<NinjaDTO>> listarNinjas() {
+        List<NinjaDTO> ninjas = ninjaService.listarNinjas();
+        return ResponseEntity.ok(ninjas);
     }
 
     @GetMapping("/listar/{id}")
-    public NinjaDTO listarNinjasPorId(@PathVariable Long id) {
-        return ninjaService.listarNinjasPorId(id);
+    public ResponseEntity<?> listarNinjasPorId(@PathVariable Long id) {
+
+        NinjaDTO ninja = ninjaService.listarNinjasPorId(id);
+
+        if (ninja != null) {
+            return ResponseEntity.ok(ninja);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja com o id: " + id + "não encontrado.");
+        }
+
     }
 
     @PutMapping("/alterar/{id}")
-    public NinjaDTO alterarNinjaPorId (@PathVariable Long id, @RequestBody NinjaDTO ninjaAtualizado) {
-        return ninjaService.atualizarNinja(id, ninjaAtualizado);
+    public ResponseEntity<?> alterarNinjaPorId (@PathVariable Long id, @RequestBody NinjaDTO ninjaAtualizado) {
+        NinjaDTO ninja = ninjaService.atualizarNinja(id, ninjaAtualizado);
+
+        if (ninja != null) {
+            return ResponseEntity.ok(ninja);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja ID: " + id + " não encontrado.");
+        }
+
     }
 
     @DeleteMapping("/deletar/{id}")
-    public void deletarNinjaPorId (@PathVariable Long id) {
-        ninjaService.deletetarNinja(id);
+    public ResponseEntity<String> deletarNinjaPorId (@PathVariable Long id) {
+
+        if (ninjaService.listarNinjasPorId(id) != null) {
+            ninjaService.deletetarNinja(id);
+            return ResponseEntity.ok("Ninja com o ID: " + id + "deletado com sucesso.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("O ninja com o ID: " + id + "não encontrado.");
+        }
+
     }
 
 }
